deleted surname. added dialog for deleting finance operations and patterns

+++ app/src/main/java/com/example/myfinance/Fragments/SettingsFragment.java
@@ -12,6 +12,7 @@ import android.widget.Toast;
 import androidx.activity.OnBackPressedCallback;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatDelegate;
 import androidx.cardview.widget.CardView;
 import androidx.fragment.app.Fragment;
@@ -97,11 +98,32 @@ public class SettingsFragment extends Fragment {
         setupSpinnerListeners();
 
         deleteAllPatterns.setOnClickListener(v -> {
-            deletePatterns();
+            new AlertDialog.Builder(requireContext())
+                    .setTitle("Удалить")
+                    .setMessage("Вы уверены, что хотите удалить Все шаблоны?")
+                    .setPositiveButton("Подтвердить", (dialog, which) -> {
+                        deletePatterns();
+                    })
+                    .setNegativeButton("Отмена", (dialog, which) -> {
+                        dialog.dismiss();
+                    })
+                    .show();
+
         });
 
         deleteAllFinances.setOnClickListener(v -> {
-            deleteFinances();
+            new AlertDialog.Builder(requireContext())
+                    .setTitle("Удалить")
+                    .setMessage("Вы уверены, что хотите удалить Все финансовые записи?")
+                    .setPositiveButton("Подтвердить", (dialog, which) -> {
+                        deleteFinances();
+                    })
+                    .setNegativeButton("Отмена", (dialog, which) -> {
+                        dialog.dismiss();
+                    })
+                    .show();
+
+
         });
     }
